move openslide to a new-style class

+++ libvips/include/vips/internal.h
@@ -294,7 +294,7 @@ void vips__init_wrap7_classes( void );
 void vips_arithmetic_operation_init( void );
 void vips_conversion_operation_init( void );
 
-size_t vips__parse_size( const char *size_string );
+guint64 vips__parse_size( const char *size_string );
 
 IMAGE *vips__deprecated_open_read( const char *filename );
 IMAGE *vips__deprecated_open_write( const char *filename );
